Move result types into repository file
